spec updates
